Split trading_bot class form main
